+++ integration-tests/common/gauntlet_plus_plus_common.go
@@ -7,7 +7,7 @@ import (
 	"os"
 
 	"github.com/smartcontractkit/chainlink-starknet/integration-tests/utils"
-	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/chainlink/txm"
+	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/starknet/devnet"
 )
 
 func (m *OCRv2TestState) fundNodesWithGPP() ([]string, error) {
@@ -40,12 +40,12 @@ func (m *OCRv2TestState) fundNodesWithGPP() ([]string, error) {
 	} else {
 		const fundAmount = uint64(900_000_000_000_000_000)
 		for _, key := range nAccounts {
-			res, err := txm.DevnetMint(m.Common.RPCDetails.RPCL2External, key, fundAmount, "WEI")
+			res, err := devnet.DevnetMint(m.Common.RPCDetails.RPCL2External, key, fundAmount, "WEI")
 			if err != nil {
 				return nil, err
 			}
 			l.Info().Msg(fmt.Sprintf("Funding account (WEI): %s", res))
-			res, err = txm.DevnetMint(m.Common.RPCDetails.RPCL2External, key, fundAmount, m.Common.ChainDetails.TokenName)
+			res, err = devnet.DevnetMint(m.Common.RPCDetails.RPCL2External, key, fundAmount, m.Common.ChainDetails.TokenName)
 			if err != nil {
 				return nil, err
 			}
