+++ integration-tests/common/test_common.go
@@ -33,8 +33,8 @@ import (
 	"github.com/smartcontractkit/chainlink-starknet/ops"
 	"github.com/smartcontractkit/chainlink-starknet/ops/gauntlet"
 	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/chainlink/ocr2"
-	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/chainlink/txm"
 	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/starknet"
+	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/starknet/devnet"
 )
 
 var (
@@ -239,7 +239,7 @@ func (m *OCRv2TestState) DeployCluster() {
 	// If we are using devnet fetch the default keys
 	if *m.Common.TestConfig.Common.Network == "localnet" {
 		// fetch predeployed account 0 to use as funder
-		accounts, err := txm.FetchDevnetAccounts(m.Common.RPCDetails.RPCL2External)
+		accounts, err := devnet.FetchDevnetAccounts(m.Common.RPCDetails.RPCL2External)
 		require.NoError(m.TestConfig.T, err)
 		account := accounts[0]
 		m.Account.Account = account.Address
