+++ relayer/pkg/starknet/devnet/devnet_test.go
@@ -1,4 +1,4 @@
-package txm
+package devnet
 
 import (
 	"net/http"
